backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
psycopg[binary]==3.1.18
pydantic==2.6.1
python-dotenv==1.0.1

langchain==0.2.7
langchain-community==0.2.7
langchain-anthropic==0.1.20
langchain-openai==0.1.14
chromadb==0.5.3
sentence-transformers==2.7.0
pymupdf==1.24.2
docx2txt==0.8
python-docx==1.1.2
rank-bm25==0.2.2

crewai==0.41.0
litellm==1.35.7
